backend/requirements.txt
# Google Drive API Client & Authentication
pydrive2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai

# Environment variables
python-dotenv

# Retries Logic
tenacity

# Requirements for ParserAgent
pdfplumber

# Google Cloud Vision API for OCR detection
google-cloud-vision

# Requirements for Excel Agent
pandas
openpyxl

# Flask Api 
flask
flask-cors

# Adding langchain imports
langchain
langchain-google-genai
langchain-core